add: serialization utilities for OS hints and contract classes
