+++ crates/core/generate-pie/src/utils/mod.rs
@@ -1,7 +1,9 @@
 pub mod commitment;
 pub mod misc;
 pub mod rpc;
+pub mod serialization;
 
 pub use commitment::*;
 pub use misc::*;
 pub use rpc::*;
+pub use serialization::*;
